udp,vxlan,ip,server: Build options and split files
After this modification compilation will be dependent on OS.
It will produce windows and linux specific versions

This is needed for further work on windows.
