+++ backend/vxlan/device.go
@@ -1,3 +1,5 @@
+// +build linux !windows
+
 // Copyright 2015 flannel authors
 //
 // Licensed under the Apache License, Version 2.0 (the "License");
